+++ perun-core/src/main/java/cz/metacentrum/perun/core/impl/SearcherImpl.java
@@ -15,6 +15,7 @@ import javax.sql.DataSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.jdbc.core.JdbcPerunTemplate;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 
 import cz.metacentrum.perun.core.api.BeansUtils;
